Rework inputs system: no more CA derivations :(
CA derivations are good in theory but too many tools have issues with
them, and you are forced to download more stuff to get output paths.

We also take this opportunity to use nixpkgs' fetchers for inputs (other
than nixpkgs). This should cache them better, but with the downside of
recompiling everything when an input changes whatsoever :(

Local overrides now are done with the nix path: `-I myinput=/a/path`
These should still work with finer grained change detection, but are not
equivalent to the npin version even if identical contents :(